backend/requirements.txt
fastapi
uvicorn
databases
sqlalchemy
pydantic[dotenv]
bcrypt
pyjwt
asyncpg
psycopg2
python-multipart
aiofiles